Google Cloud Docker Deploy: Try new command
The Cloud console was createing a command to build the docker image.

We were staying in the root directory to do things but it seems like
we really want to move into the Docker directory to build things.

This cleans up some other attempts to figure out why we build but
don't get any response.